Reducing publication post-processing seems to work.

+++ MillimanAccessPortal/MillimanAccessPortal/Services/QueuedGoLiveTaskHostedService.cs
@@ -194,7 +194,119 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
 
         using (IDbContextTransaction Txn = dbContext.Database.BeginTransaction())
         {
-            // 1 Move new master content and related files (not reduced content) into live file names,
+            //1 Update db:
+            //1.1  ContentPublicationRequest.Status
+            var previousLiveRequests = dbContext.ContentPublicationRequest
+                .Where(r => r.RootContentItemId == publicationRequest.RootContentItemId)
+                .Where(r => r.RequestStatus == PublicationStatus.Confirmed);
+            foreach (ContentPublicationRequest PreviousLiveRequest in previousLiveRequests)
+            {
+                PreviousLiveRequest.RequestStatus = PublicationStatus.Replaced;
+                dbContext.ContentPublicationRequest.Update(PreviousLiveRequest);
+            }
+            publicationRequest.RequestStatus = PublicationStatus.Confirmed;
+
+            //1.2  ContentReductionTask.Status
+            var previousLiveTasks = dbContext.ContentReductionTask
+                .Where(r => r.SelectionGroup.RootContentItemId == publicationRequest.RootContentItemId)
+                .Where(r => r.ReductionStatus == ReductionStatusEnum.Live);
+            foreach (ContentReductionTask PreviousLiveTask in previousLiveTasks)
+            {
+                PreviousLiveTask.ReductionStatus = ReductionStatusEnum.Replaced;
+                dbContext.ContentReductionTask.Update(PreviousLiveTask);
+            }
+            relatedReductionTasks.ForEach(t => t.ReductionStatus = ReductionStatusEnum.Live);
+
+            //1.3  HierarchyFieldValue due to hierarchy changes
+            //1.3.1  If this is first publication for this root content item, add the fields to db
+            //       and to LiveHierarchy to help identify all values as new
+            if (LiveHierarchy.Fields.Count == 0)
+            {  // This must be first time publication, need to insert the fields.  Values are handled below
+                NewHierarchy.Fields.ForEach(f =>
+                {
+                    HierarchyField NewField = new HierarchyField
+                    {
+                        FieldName = f.FieldName,
+                        FieldDisplayName = f.DisplayName,
+                        RootContentItemId = publicationRequest.RootContentItemId,
+                        FieldDelimiter = f.ValueDelimiter,
+                        StructureType = f.StructureType,
+                    };
+                    dbContext.HierarchyField.Add(NewField);
+                    dbContext.SaveChanges();
+
+                    LiveHierarchy.Fields.Add(new ReductionField<ReductionFieldValue>
+                    {
+                        Id = NewField.Id,  // Id is assigned during dbContext.SaveChanges() above
+                        FieldName = NewField.FieldName,
+                        DisplayName = NewField.FieldDisplayName,
+                        StructureType = NewField.StructureType,
+                        ValueDelimiter = NewField.FieldDelimiter,
+                        Values = new List<ReductionFieldValue>(),
+                    });
+                });
+            }
+            //1.3.2  Add/Remove field values based on value list differences between new/old
+            foreach (var NewHierarchyField in NewHierarchy.Fields)
+            {
+                ReductionField<ReductionFieldValue> MatchingLiveField = LiveHierarchy.Fields
+                    .Single(f => f.FieldName == NewHierarchyField.FieldName);
+
+                List<string> NewHierarchyFieldValueList = NewHierarchyField.Values.Select(v => v.Value).ToList();
+                List<string> LiveHierarchyFieldValueList = MatchingLiveField.Values.Select(v => v.Value).ToList();
+
+                // Insert new values
+                foreach (string NewValue in NewHierarchyFieldValueList.Except(LiveHierarchyFieldValueList))
+                {
+                    dbContext.HierarchyFieldValue.Add(new HierarchyFieldValue
+                    {
+                        HierarchyFieldId = MatchingLiveField.Id,
+                        Value = NewValue,
+                    });
+                }
+
+                // Delete removed values
+                foreach (string RemovedValue in LiveHierarchyFieldValueList.Except(NewHierarchyFieldValueList))
+                {
+                    HierarchyFieldValue ObsoleteRecord = dbContext.HierarchyFieldValue
+                        .Single(v =>
+                            v.HierarchyFieldId == MatchingLiveField.Id
+                            && v.Value == RemovedValue);
+                    dbContext.HierarchyFieldValue.Remove(ObsoleteRecord);
+                }
+            }
+            dbContext.SaveChanges();
+
+            //1.4  Update SelectionGroup SelectedHierarchyFieldValueList due to hierarchy changes
+            List<Guid> AllRemainingFieldValues = dbContext.HierarchyFieldValue
+                .Where(v => v.HierarchyField.RootContentItemId == publicationRequest.RootContentItemId)
+                .Select(v => v.Id)
+                .ToList();
+            var reducingSelectionGroups = dbContext.SelectionGroup
+                .Where(g => g.RootContentItemId == publicationRequest.RootContentItemId && !g.IsMaster);
+            foreach (SelectionGroup Group in reducingSelectionGroups)
+            {
+                Group.SelectedHierarchyFieldValueList = Group.SelectedHierarchyFieldValueList
+                    .Intersect(AllRemainingFieldValues).ToArray();
+            }
+
+            // Perform any content type dependent follow up processing
+            switch (publicationRequest.RootContentItem.ContentType.TypeEnum)
+            {
+                case ContentTypeEnum.Qlikview:
+                    await new QlikviewLibApi()
+                        .AuthorizeUserDocumentsInFolder(
+                            goLiveViewModel.RootContentItemId.ToString(), qlikviewConfig);
+                    break;
+
+                case ContentTypeEnum.Html:
+                case ContentTypeEnum.Pdf:
+                case ContentTypeEnum.FileDownload:
+                default:
+                    break;
+            }
+
+            // 2 Move new master content and related files (not reduced content) into live file names,
             //   removing any existing copies of previous version
             List<ContentRelatedFile> UpdatedContentFilesList = publicationRequest.RootContentItem.ContentFilesList;
             foreach (ContentRelatedFile Crf in publicationRequest.LiveReadyFilesObj)
@@ -202,7 +314,9 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                 // This assignment defines the live file name
                 string TargetFileName = ContentTypeSpecificApiBase.GenerateContentFileName(
                     Crf.FilePurpose, Path.GetExtension(Crf.FullPath), goLiveViewModel.RootContentItemId);
-                string TargetFilePath = Path.Combine(Path.GetDirectoryName(Crf.FullPath), TargetFileName);
+                string TargetFilePath = Path.Combine(configuration.GetSection("Storage")["ContentItemRootPath"],
+                                                     goLiveViewModel.RootContentItemId.ToString(),
+                                                     TargetFileName);
 
                 // Move any existing file to backed up name
                 if (File.Exists(TargetFilePath))
@@ -275,7 +389,7 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
             }
             publicationRequest.RootContentItem.ContentFilesList = UpdatedContentFilesList;
 
-            // 2 Rename reduced content files to live names
+            // 3 Rename reduced content files to live names
             foreach (var ThisTask in relatedReductionTasks.Where(t => t.SelectionGroupId.HasValue).Where(t => !t.SelectionGroup.IsMaster))
             {
                 // This assignment defines the live file name for any reduced content file
@@ -317,121 +431,12 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
 
                 if (!isInactive)
                 {
-                    File.Copy(ThisTask.ResultFilePath, TargetFilePath);
-                    FilesToDelete.Add(ThisTask.ResultFilePath);
+                    File.Move(ThisTask.ResultFilePath, TargetFilePath);
                 }
-            }
-
-            //3 Update db:
-            //3.1  ContentPublicationRequest.Status
-            var previousLiveRequests = dbContext.ContentPublicationRequest
-                .Where(r => r.RootContentItemId == publicationRequest.RootContentItemId)
-                .Where(r => r.RequestStatus == PublicationStatus.Confirmed);
-            foreach (ContentPublicationRequest PreviousLiveRequest in previousLiveRequests)
-            {
-                PreviousLiveRequest.RequestStatus = PublicationStatus.Replaced;
-                dbContext.ContentPublicationRequest.Update(PreviousLiveRequest);
-            }
-            publicationRequest.RequestStatus = PublicationStatus.Confirmed;
-
-            //3.2  ContentReductionTask.Status
-            var previousLiveTasks = dbContext.ContentReductionTask
-                .Where(r => r.SelectionGroup.RootContentItemId == publicationRequest.RootContentItemId)
-                .Where(r => r.ReductionStatus == ReductionStatusEnum.Live);
-            foreach (ContentReductionTask PreviousLiveTask in previousLiveTasks)
-            {
-                PreviousLiveTask.ReductionStatus = ReductionStatusEnum.Replaced;
-                dbContext.ContentReductionTask.Update(PreviousLiveTask);
-            }
-            relatedReductionTasks.ForEach(t => t.ReductionStatus = ReductionStatusEnum.Live);
-
-            //3.3  HierarchyFieldValue due to hierarchy changes
-            //3.3.1  If this is first publication for this root content item, add the fields to db
-            //       and to LiveHierarchy to help identify all values as new
-            if (LiveHierarchy.Fields.Count == 0)
-            {  // This must be first time publication, need to insert the fields.  Values are handled below
-                NewHierarchy.Fields.ForEach(f =>
+                else   // TODO path is null or empty?, remove this, it's throwing during File.Delete
                 {
-                    HierarchyField NewField = new HierarchyField
-                    {
-                        FieldName = f.FieldName,
-                        FieldDisplayName = f.DisplayName,
-                        RootContentItemId = publicationRequest.RootContentItemId,
-                        FieldDelimiter = f.ValueDelimiter,
-                        StructureType = f.StructureType,
-                    };
-                    dbContext.HierarchyField.Add(NewField);
-                    dbContext.SaveChanges();
-
-                    LiveHierarchy.Fields.Add(new ReductionField<ReductionFieldValue>
-                    {
-                        Id = NewField.Id,  // Id is assigned during dbContext.SaveChanges() above
-                        FieldName = NewField.FieldName,
-                        DisplayName = NewField.FieldDisplayName,
-                        StructureType = NewField.StructureType,
-                        ValueDelimiter = NewField.FieldDelimiter,
-                        Values = new List<ReductionFieldValue>(),
-                    });
-                });
-            }
-            //3.3.2  Add/Remove field values based on value list differences between new/old
-            foreach (var NewHierarchyField in NewHierarchy.Fields)
-            {
-                ReductionField<ReductionFieldValue> MatchingLiveField = LiveHierarchy.Fields
-                    .Single(f => f.FieldName == NewHierarchyField.FieldName);
-
-                List<string> NewHierarchyFieldValueList = NewHierarchyField.Values.Select(v => v.Value).ToList();
-                List<string> LiveHierarchyFieldValueList = MatchingLiveField.Values.Select(v => v.Value).ToList();
-
-                // Insert new values
-                foreach (string NewValue in NewHierarchyFieldValueList.Except(LiveHierarchyFieldValueList))
-                {
-                    dbContext.HierarchyFieldValue.Add(new HierarchyFieldValue
-                    {
-                        HierarchyFieldId = MatchingLiveField.Id,
-                        Value = NewValue,
-                    });
+                    //FilesToDelete.Add(ThisTask.ResultFilePath);  
                 }
-
-                // Delete removed values
-                foreach (string RemovedValue in LiveHierarchyFieldValueList.Except(NewHierarchyFieldValueList))
-                {
-                    HierarchyFieldValue ObsoleteRecord = dbContext.HierarchyFieldValue
-                        .Single(v =>
-                            v.HierarchyFieldId == MatchingLiveField.Id
-                            && v.Value == RemovedValue);
-                    dbContext.HierarchyFieldValue.Remove(ObsoleteRecord);
-                }
-            }
-            dbContext.SaveChanges();
-
-            //3.4  Update SelectionGroup SelectedHierarchyFieldValueList due to hierarchy changes
-            List<Guid> AllRemainingFieldValues = dbContext.HierarchyFieldValue
-                .Where(v => v.HierarchyField.RootContentItemId == publicationRequest.RootContentItemId)
-                .Select(v => v.Id)
-                .ToList();
-            var reducingSelectionGroups = dbContext.SelectionGroup
-                .Where(g => g.RootContentItemId == publicationRequest.RootContentItemId && !g.IsMaster);
-            foreach (SelectionGroup Group in reducingSelectionGroups)
-            {
-                Group.SelectedHierarchyFieldValueList = Group.SelectedHierarchyFieldValueList
-                    .Intersect(AllRemainingFieldValues).ToArray();
-            }
-
-            // Perform any content type dependent follow up processing
-            switch (publicationRequest.RootContentItem.ContentType.TypeEnum)
-            {
-                case ContentTypeEnum.Qlikview:
-                    await new QlikviewLibApi()
-                        .AuthorizeUserDocumentsInFolder(
-                            goLiveViewModel.RootContentItemId.ToString(), qlikviewConfig);
-                    break;
-
-                case ContentTypeEnum.Html:
-                case ContentTypeEnum.Pdf:
-                case ContentTypeEnum.FileDownload:
-                default:
-                    break;
             }
 
             dbContext.SaveChanges();
@@ -447,21 +452,21 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
         // 4 Delete all temporary files
         foreach (string FileToDelete in FilesToDelete)
         {
-            try
+            if (File.Exists(FileToDelete))
             {
                 File.Delete(FileToDelete);
-            }
-            catch (DirectoryNotFoundException)
-            {
-                continue;
             }
         }
 
         if (ReductionIsInvolved)
         {
-            // Delete temporary folder of publication job (contains temporary reduced content files)
-            string PubJobTempFolder = Path.GetDirectoryName(relatedReductionTasks[0].MasterFilePath);
-            Directory.Delete(PubJobTempFolder, true);
+            string tempPublicationFolder = Path.Combine(configuration.GetSection("Storage")["ContentItemRootPath"],
+                                                        publicationRequest.RootContentItemId.ToString(),
+                                                        publicationRequest.Id.ToString());
+            if (Directory.Exists(tempPublicationFolder))
+            {
+                Directory.Delete(tempPublicationFolder, true);
+            }
         }
         #endregion
     }
